# public/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Znaidi Mahdi Starter - Open API",
        "description": "APIs documentation usin OpenAPI",
        "contact": {
            "name": "Znaidi Mahdi",
            "email": "mahdi.znaidi@esprit.tn"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "a local dev server"
        }
    ],
    "paths": {
...
        "/types/exp": {
            "get": {
                "tags": [
                    "TypeExperience"
                ],
                "summary": "List all TypeExps",
                "operationId": "indexTypeExp",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "List All TypeExps",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "List All TypeExps",
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "description": "List All TypeExps",
                                            "type": "string",
                                            "example": "typeExps_list"
                                        },
                                        "data": {
                                            "description": "List All TypeExps",
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/TypeExp"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        },
        "/types/{id}/exp": {
            "get": {
                "tags": [
                    "TypeExperience"
                ],
                "summary": "Show an existing TypeExp",
                "operationId": "showTypeExp",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of TypeExp to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Show an existing TypeExp",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "Show an existing TypeExp",
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "description": "Show an existing TypeExp",
                                            "type": "string",
                                            "example": "typeExp_details"
                                        },
                                        "data": {
                                            "description": "Show an existing TypeExp",
                                            "items": {
                                                "$ref": "#/components/schemas/TypeExp"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "TypeExperience"
                ],
                "summary": "Delete an existing TypeExp",
                "operationId": "destroyTypeExp",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of TypeExp to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Delete an existing TypeExp",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "Delete an existing TypeExp",
                                            "type": "integer",
                                            "example": 204
                                        },
                                        "message": {
                                            "description": "Delete an existing TypeExp",
                                            "type": "string",
                                            "example": "typeExp_deleted"
                                        },
                                        "data": {
                                            "description": "Delete an existing TypeExp",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "TypeExperience"
                ],
                "summary": "Update an existing TypeExp",
                "operationId": "updateTypeExps",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of equipe to return",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/requestBodies/TypeExp"
                },
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Update an existing TypeExp",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "Update an existing TypeExp",
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "description": "Update an existing TypeExp",
                                            "type": "string",
                                            "example": "typeExp_updated"
                                        },
                                        "data": {
                                            "description": "Update an existing TypeExp",
                                            "items": {
                                                "$ref": "#/components/schemas/TypeExp"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        },
        "/types/exps": {
            "post": {
                "tags": [
                    "TypeExperience"
                ],
                "summary": "Create a new TypeExp",
                "operationId": "storeTypeExp",
                "requestBody": {
                    "$ref": "#/components/requestBodies/TypeExp"
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "description": "Create a new TypeExp",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "code": {
                                            "description": "Create a new TypeExp",
                                            "type": "integer",
                                            "example": 201
                                        },
                                        "message": {
                                            "description": "Create a new TypeExp",
                                            "type": "string",
                                            "example": "typeExp_created"
                                        },
                                        "data": {
                                            "description": "Create a new TypeExp",
                                            "items": {
                                                "$ref": "#/components/schemas/TypeExp"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "Password Based": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Experience": {
                "title": "Experience",
                "description": "Experience model",
                "required": [
                    "intitule"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "Experience ID",
                        "type": "integer",
                        "format": "int"
                    },
                    "intitule": {
                        "title": "intitule",
                        "description": "Experience intitule",
                        "type": "string",
                        "format": "string"
                    },
                    "date_debut": {
                        "title": "date_debut",
                        "description": "Experience date_debut",
                        "type": "string",
                        "format": "string"
                    },
                    "date_fin": {
                        "title": "date_fin",
                        "description": "Experience date_fin",
                        "type": "string",
                        "format": "string"
                    },
                    "user": {
                        "title": "users",
                        "description": "Experience user",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/User"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Experience"
                }
            },
...
            "TypeExp": {
                "title": "TypeExp",
                "description": "TypeExp model",
                "required": [
                    "libelle"
                ],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "TypeExp ID",
                        "type": "integer",
                        "format": "int"
                    },
                    "libelle": {
                        "title": "string",
                        "description": "TypeExp libelle",
                        "type": "integer",
                        "format": "string"
                    },
                    "experiences": {
                        "title": "experiences",
                        "description": "TypeExp experiences",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Experience"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TypeExp"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "User id",
                        "type": "integer",
                        "format": "integer"
                    },
                    "email": {
                        "title": "email",
                        "description": "User email",
                        "type": "string",
                        "format": "string"
                    },
                    "matricule_hr": {
                        "title": "matricule_hr",
                        "description": "User matricule_hr",
                        "type": "string",
                        "format": "string"
                    },
                    "nom": {
                        "title": "nom",
                        "description": "User nom",
                        "type": "string",
                        "format": "string"
                    },
                    "prenom": {
                        "title": "prenom",
                        "description": "User prenom",
                        "type": "string",
                        "format": "string"
                    },
                    "adresse": {
                        "title": "adresse",
                        "description": "User adresse",
                        "type": "string",
                        "format": "string"
                    },
                    "num_telephone": {
                        "title": "numéro_téléphone",
                        "description": "User numéro_telephone",
                        "type": "string",
                        "format": "string"
                    },
                    "cin_passport": {
                        "title": "cin_passport",
                        "description": "User cin_passport",
                        "type": "string",
                        "format": "string"
                    },
                    "etat_civil": {
                        "title": "etat_civil",
                        "description": "User etat_civil",
                        "type": "string",
                        "format": "string"
                    },
                    "image": {
                        "title": "image",
                        "description": "User image",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            },
...
            "TypeExpRequestBody": {
                "title": "TypeExpRequestBody",
                "description": "TypeExp model",
                "properties": {
                    "libelle": {
                        "title": "libelle",
                        "description": "TypeExp libelle",
                        "type": "string",
                        "format": "string"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "TypeExp"
                }
            },
...
        },
        "requestBodies": {
...
            "TypeExp": {
                "description": "TypeExp object that needs to be added to the store",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/TypeExpRequestBody"
                        }
                    }
                }
            },
...
        },
        "securitySchemes": {
            "Password Based": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}